platform/backend/requirements.txt
Django==3.2.23 
asgiref==3.7.2
sqlparse==0.4.4
psycopg2-binary==2.9.9
djangorestframework==3.14.0
docker==6.1.3
django-cors-headers==4.3.1
python-dotenv==1.0.1
uuid==1.30
boto3==1.34.114
django-storages==1.14.3
